+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.3.0
 
 * feat: (fxJob) allows to run fxJob locally
+* feat: (fxJob) auto-set launchLocally to true if env var JENKINS_LOCAL exists
 * feat: (fxJob) disables cleanWs when run locally
 * feat: (dockerRunCommand) allows to change docker run data basepath
 * feat: (dockerRunCommand) allows to run docker commands as deamon